Реальная скорость копирования считается нормально. Нужно переделать её на милли/наносекунды.

+++ copy_between_streams_with_speed/src/main/java/com/github/aakumykov/copy_between_streams_with_speed/millisecondsToDHMSN.kt
@@ -4,7 +4,7 @@ import java.util.Locale
 import kotlin.math.roundToLong
 import kotlin.time.Duration.Companion.milliseconds
 
-fun millisecondsToDHMSN(time: Long): String = time.milliseconds.toComponents { days, hours, minutes, seconds, nanoseconds ->
+fun millisecondsToDHMSN(time: Double): String = time.milliseconds.toComponents { days, hours, minutes, seconds, nanoseconds ->
     String.format(
         Locale.getDefault(),
         "%02dд, %02dч, %02dм, %02dс, %02dмc",
